Make `resolvelib.resolvers` a sub-package
This makes the `Resolver` object live in a diffferent module than the
`AbstractResolver`, clarifying their separation and more cleanly
accomodating additional resolver implementations.
